Use bracket_ instead of (flip finally) to manage adding and deleting POs from networks.
This solves a _very_ subtle problem that was causing segfaults.  An exception was being
raised during the process of adding POs, because some arguments contained unevaluated 'error'
calls.  This would cause the 'finally' to execute the code for deleting POs.  As the network
was in an inconsistent state after the exception, this would segfault.

With bracket_, the exception handler is not installed until after the process of adding POs
is completed.  Now, if an exception occurs in this process we might still be in an inconsistent
state especially if some POs are successfully added but others are not :-(
However, at least the immediate segfault problem is solved.