Newer clang problems on OSX with c2hs
c2hs has problems with things that use `__has_feature`, so in the same
vein as the previous use of `-U__has_attribute`, this also undefines the
`__has_feature` macro. It would be nice if we could guarantee that these
changes were only being used when running c2hs, but there doesn't seem
to be a way to differentiate with cabal.