Fall back on copying GIAs for write, cec
The Data.ABC.GIA.withNetworkPtr function is implemented in two ways. One
copies the network.  The other attempts to avoid copying the network by
modifying it in place and restoring it. The latter seemed questionable
with old ABC versions, but causes explicit assertion failures with newer
versions, so this commit disables it. If the copying version turns out
to be too expensive, we can revisit this with a more reliable approach.
It should only affect writing AIG files and doing equivalence checks.